Actualizacion 27/9/2020 creado el sistema de roles

+++ database/migrations/2019_07_14_174332_create_perfumes_table.php
@@ -20,7 +20,9 @@ class CreatePerfumesTable extends Migration
             $table->text('extract');
             $table->decimal('price', 5, 2);
             $table->string('imgPerfume');
-            $table->integer('category_id')->unsigned();
+            
+            $table->bigInteger('category_id')->unsigned();
+            
             $table->foreign('category_id')
                   ->references('id')
                   ->on('categories')
